Split test_config.py: extract quota, tracing, and branching config tests
Extract TestQuotaGuardConfig, TestLinuxTracingConfig, TestBranchingConfig,
and TestReleaseReadinessConfig into focused files. test_config.py shrinks
from 891 to 605 lines. All new files carry pytestmark layer("config") + small.
