+++ tests/config/test_config.py
@@ -337,142 +337,6 @@ class TestRunSkillConfigExitAfterStopDelay:
         assert "exit_after_stop_delay_ms" in names
 
 
-class TestQuotaGuardConfig:
-    def test_default_enabled(self):
-        import pytest
-
-        config = AutomationConfig()
-        assert config.quota_guard.enabled is True
-        assert config.quota_guard.short_window_threshold == pytest.approx(85.0)
-        assert config.quota_guard.long_window_threshold == pytest.approx(95.0)
-        assert config.quota_guard.buffer_seconds == 60
-        assert config.quota_guard.cache_max_age == 300
-
-    def test_load_quota_guard_from_yaml(self, tmp_path):
-        import pytest
-
-        config_dir = tmp_path / ".autoskillit"
-        config_dir.mkdir()
-        (config_dir / "config.yaml").write_text(
-            yaml.dump(
-                {
-                    "quota_guard": {
-                        "enabled": True,
-                        "short_window_threshold": 85.0,
-                        "long_window_threshold": 95.0,
-                    }
-                }
-            )
-        )
-        config = load_config(tmp_path)
-        assert config.quota_guard.enabled is True
-        assert config.quota_guard.short_window_threshold == pytest.approx(85.0)
-        assert config.quota_guard.long_window_threshold == pytest.approx(95.0)
-        # Unspecified fields keep defaults
-        assert config.quota_guard.buffer_seconds == 60
-
-    def test_short_window_threshold_defaults_to_85(self):
-        from autoskillit.config.settings import QuotaGuardConfig
-
-        assert QuotaGuardConfig().short_window_threshold == 85.0
-
-    def test_long_window_threshold_defaults_to_95(self):
-        from autoskillit.config.settings import QuotaGuardConfig
-
-        assert QuotaGuardConfig().long_window_threshold == 95.0
-
-    def test_long_window_patterns_default(self):
-        from autoskillit.config.settings import QuotaGuardConfig
-
-        defaults = QuotaGuardConfig().long_window_patterns
-        assert "seven_day" in defaults, (
-            f"Default long_window_patterns {defaults!r} must include 'seven_day' "
-            "(the actual Anthropic API key for the 7-day budget)"
-        )
-
-    def test_threshold_field_removed(self):
-        import dataclasses
-
-        from autoskillit.config.settings import QuotaGuardConfig
-
-        names = {f.name for f in dataclasses.fields(QuotaGuardConfig)}
-        assert "threshold" not in names
-
-    def test_quota_guard_yaml_round_trip_per_window(self, tmp_path):
-        import pytest
-
-        config_dir = tmp_path / ".autoskillit"
-        config_dir.mkdir()
-        (config_dir / "config.yaml").write_text(
-            yaml.dump(
-                {
-                    "quota_guard": {
-                        "short_window_threshold": 85.0,
-                        "long_window_threshold": 95.0,
-                        "long_window_patterns": ["weekly", "sonnet", "opus"],
-                    }
-                }
-            )
-        )
-        config = load_config(tmp_path)
-        assert config.quota_guard.short_window_threshold == pytest.approx(85.0)
-        assert config.quota_guard.long_window_threshold == pytest.approx(95.0)
-        assert config.quota_guard.long_window_patterns == ["weekly", "sonnet", "opus"]
-
-    def test_quota_guard_config_has_cache_refresh_interval(self):
-        """QG_C3: QuotaGuardConfig has cache_refresh_interval defaulting to 240."""
-        from autoskillit.config.settings import QuotaGuardConfig
-
-        config = QuotaGuardConfig()
-        assert config.cache_refresh_interval == 240
-
-    def test_defaults_yaml_has_cache_refresh_interval(self):
-        """QG_C4: defaults.yaml defines quota_guard.cache_refresh_interval < cache_max_age."""
-        from autoskillit.core.paths import pkg_root
-
-        defaults = yaml.safe_load((pkg_root() / "config" / "defaults.yaml").read_text())
-        assert "cache_refresh_interval" in defaults["quota_guard"]
-        interval = defaults["quota_guard"]["cache_refresh_interval"]
-        max_age = defaults["quota_guard"]["cache_max_age"]
-        assert interval < max_age, (
-            f"cache_refresh_interval ({interval}) must be < cache_max_age ({max_age}); "
-            "otherwise the loop arrives after the cache has already expired"
-        )
-
-    def test_quota_guard_per_window_enabled_defaults_true(self):
-        """Test 14: QuotaGuardConfig() defaults both per-window flags to True."""
-        from autoskillit.config.settings import QuotaGuardConfig
-
-        config = QuotaGuardConfig()
-        assert config.short_window_enabled is True
-        assert config.long_window_enabled is True
-
-    def test_quota_guard_per_window_enabled_yaml_round_trip(self, tmp_path):
-        """Test 15: per-window enabled flags survive a YAML round-trip."""
-        config_dir = tmp_path / ".autoskillit"
-        config_dir.mkdir()
-        (config_dir / "config.yaml").write_text(
-            yaml.dump({"quota_guard": {"short_window_enabled": False}})
-        )
-        config = load_config(tmp_path)
-        assert config.quota_guard.short_window_enabled is False
-        assert config.quota_guard.long_window_enabled is True
-
-    def test_defaults_yaml_has_per_window_enabled_keys(self):
-        """Test 16: defaults.yaml has both per-window enabled keys set to True."""
-        from autoskillit.core.paths import pkg_root
-
-        defaults = yaml.safe_load((pkg_root() / "config" / "defaults.yaml").read_text())
-        assert defaults["quota_guard"]["short_window_enabled"] is True
-        assert defaults["quota_guard"]["long_window_enabled"] is True
-
-    def test_quota_guard_env_var_override_short_window_enabled(self, monkeypatch, tmp_path):
-        """Test 17: AUTOSKILLIT_QUOTA_GUARD__SHORT_WINDOW_ENABLED=false is cast to bool False."""
-        monkeypatch.setenv("AUTOSKILLIT_QUOTA_GUARD__SHORT_WINDOW_ENABLED", "false")
-        config = load_config(tmp_path)
-        assert config.quota_guard.short_window_enabled is False
-
-
 class TestLoggingConfig:
     """LoggingConfig dataclass and YAML loading."""
 
@@ -538,69 +402,6 @@ class TestLoggingConfig:
 
         names = {f.name for f in dc_fields(LoggingConfig)}
         assert names == {"level", "json_output"}
-
-
-class TestLinuxTracingConfig:
-    """LinuxTracingConfig dataclass and YAML loading."""
-
-    def test_linux_tracing_config_defaults(self, tmp_path):
-        """LT_C1: LinuxTracingConfig defaults: enabled, 5s interval, empty log_dir."""
-        cfg = load_config(tmp_path)
-        assert cfg.linux_tracing.enabled is True
-        assert cfg.linux_tracing.proc_interval == 5.0
-        assert cfg.linux_tracing.log_dir == ""
-
-    def test_linux_tracing_config_from_yaml(self, tmp_path):
-        """LT_C2: LinuxTracingConfig reads from project config."""
-        config_dir = tmp_path / ".autoskillit"
-        config_dir.mkdir()
-        (config_dir / "config.yaml").write_text(
-            "linux_tracing:\n  enabled: true\n  proc_interval: 2.0\n  log_dir: /custom/logs\n"
-        )
-        cfg = load_config(tmp_path)
-        assert cfg.linux_tracing.enabled is True
-        assert cfg.linux_tracing.proc_interval == 2.0
-        assert cfg.linux_tracing.log_dir == "/custom/logs"
-
-    def test_automation_config_has_linux_tracing_field(self):
-        """LT_C3: AutomationConfig has linux_tracing sub-config."""
-        cfg = AutomationConfig()
-        assert cfg.linux_tracing.enabled is True
-        assert cfg.linux_tracing.proc_interval == 5.0
-        assert cfg.linux_tracing.log_dir == ""
-
-    def test_linux_tracing_config_fields(self):
-        """LT_C4: LinuxTracingConfig has exactly the expected fields."""
-        from dataclasses import fields as dc_fields
-
-        from autoskillit.config.settings import LinuxTracingConfig
-
-        names = {f.name for f in dc_fields(LinuxTracingConfig)}
-        assert names == {"enabled", "proc_interval", "log_dir", "tmpfs_path", "max_sessions"}
-
-    def test_linux_tracing_max_sessions_default(self):
-        """LT_C5: LinuxTracingConfig includes max_sessions with default 2000."""
-        import dataclasses
-
-        from autoskillit.config.settings import LinuxTracingConfig
-
-        cfg = LinuxTracingConfig(tmpfs_path="/tmp/test")
-        assert cfg.max_sessions == 2000
-        assert "max_sessions" in {f.name for f in dataclasses.fields(LinuxTracingConfig)}
-
-    def test_linux_tracing_max_sessions_yaml_roundtrip(self, tmp_path):
-        """LT_C6: max_sessions survives YAML round-trip."""
-        (tmp_path / ".autoskillit").mkdir()
-        (tmp_path / ".autoskillit" / "config.yaml").write_text(
-            "linux_tracing:\n  max_sessions: 750\n"
-        )
-        cfg = load_config(tmp_path)
-        assert cfg.linux_tracing.max_sessions == 750
-
-    def test_automation_config_linux_tracing_has_max_sessions(self):
-        """LT_C3 extension: AutomationConfig().linux_tracing includes max_sessions."""
-        cfg = AutomationConfig()
-        assert cfg.linux_tracing.max_sessions == 2000
 
 
 class TestDynaconfIntegration:
@@ -780,93 +581,6 @@ def test_secrets_only_keys_covers_all_github_secret_fields() -> None:
                 f"GitHubConfig.{f.name} looks like a secret field but is missing from "
                 f"_SECRETS_ONLY_KEYS. Add 'github.{f.name}' to the frozenset in settings.py."
             )
-
-
-class TestReleaseReadinessConfig:
-    def test_branching_default_base_branch_is_main(self):
-        from autoskillit.core.io import load_yaml
-        from autoskillit.core.paths import pkg_root
-
-        defaults = load_yaml(pkg_root() / "config" / "defaults.yaml")
-        assert defaults["branching"]["default_base_branch"] == "main"
-
-
-class TestBranchingConfig:
-    def test_branching_config_default_base_branch_is_main(self) -> None:
-        """BranchingConfig must default default_base_branch to 'main'."""
-        from autoskillit.config.settings import BranchingConfig
-
-        assert BranchingConfig().default_base_branch == "main"
-
-    def test_automation_config_has_branching_field(self) -> None:
-        """AutomationConfig must expose a BranchingConfig as .branching."""
-        from autoskillit.config.settings import AutomationConfig
-
-        cfg = AutomationConfig()
-        assert cfg.branching.default_base_branch == "main"
-
-    def test_branching_config_is_overridable(self) -> None:
-        """BranchingConfig.default_base_branch must accept override values."""
-        from autoskillit.config.settings import BranchingConfig
-
-        cfg = BranchingConfig(default_base_branch="develop")
-        assert cfg.default_base_branch == "develop"
-
-    def test_branching_default_base_branch_matches_defaults_yaml(self) -> None:
-        """BranchingConfig Python default must match defaults.yaml."""
-        from autoskillit.config.settings import BranchingConfig
-        from autoskillit.core.io import load_yaml
-        from autoskillit.core.paths import pkg_root
-
-        defaults = load_yaml(pkg_root() / "config" / "defaults.yaml")
-        yaml_default = defaults["branching"]["default_base_branch"]
-        python_default = BranchingConfig().default_base_branch
-
-        assert python_default == yaml_default, (
-            f"BranchingConfig.default_base_branch Python default ({python_default!r}) "
-            f"disagrees with defaults.yaml ({yaml_default!r})"
-        )
-
-    def test_branching_config_promotion_target_defaults_to_main(self) -> None:
-        """BranchingConfig.promotion_target defaults to main (package default)."""
-        from autoskillit.config.settings import BranchingConfig
-
-        assert BranchingConfig().promotion_target == "main"
-
-    def test_automation_config_branching_promotion_target_default(self) -> None:
-        """AutomationConfig propagates promotion_target default."""
-        from autoskillit.config.settings import AutomationConfig
-
-        assert AutomationConfig().branching.promotion_target == "main"
-
-    def test_branching_config_promotion_target_overridable(self) -> None:
-        """promotion_target can be set independently of default_base_branch."""
-        from autoskillit.config.settings import BranchingConfig
-
-        cfg = BranchingConfig(default_base_branch="develop", promotion_target="main")
-        assert cfg.default_base_branch == "develop"
-        assert cfg.promotion_target == "main"
-
-    def test_branching_config_promotion_target_defaults_match_yaml(self, tmp_path) -> None:
-        """Python default for promotion_target matches defaults.yaml."""
-        from autoskillit.config import load_config
-        from autoskillit.core.io import load_yaml
-        from autoskillit.core.paths import pkg_root
-
-        loaded = load_config(tmp_path / "settings.toml")
-        defaults = load_yaml(pkg_root() / "config" / "defaults.yaml")
-        expected = defaults["branching"]["promotion_target"]
-        assert loaded.branching.promotion_target == expected
-
-    def test_branching_config_promotion_target_env_var_override(
-        self, monkeypatch, tmp_path
-    ) -> None:
-        """AUTOSKILLIT_BRANCHING__PROMOTION_TARGET env var overrides promotion_target."""
-        from autoskillit.config import load_config
-
-        monkeypatch.setenv("AUTOSKILLIT_BRANCHING__PROMOTION_TARGET", "stable")
-        cfg = load_config(tmp_path)
-        assert cfg.branching.promotion_target == "stable"
 
 
 class TestWorkspaceConfig:
